Unet_Based_DRIVEDATASET/requirements.txt
numpy==1.23.5
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
